requirements.txt
attrs==19.3.0
catboost==0.23.2
category_encoders==2.2.2
cycler==0.10.0
decorator==4.4.2
dill==0.3.2
ds-tools==0.1.3
envyaml==1.6.201229
ipykernel==5.3.1
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter_contrib_nbextensions==0.5.1
jupyter-client==6.1.5
jupyter-core==4.6.3
matplotlib==3.3.0
more-itertools==8.4.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.4
pandas==1.1.4
prometheus-client==0.8.0
prompt-toolkit==3.0.5
pyarrow==0.17.1
Pygments==2.6.1
python-dateutil==2.8.1
requests==2.24.0
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
traitlets==4.3.3
urllib3==1.25.9
widgetsnbextension==3.5.1